% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationships.R
\name{mesh_ancestors}
\alias{mesh_ancestors}
\title{Find ancestor terms}
\usage{
mesh_ancestors(mesh, db = meshdb())
}
\arguments{
\item{mesh}{Character vector of MeSH IDs or exact MeSH term labels.}

\item{db}{A \code{meshdb} object returned by \code{\link[=meshdb]{meshdb()}}.}
}
\value{
A tibble with one row per input term and ancestor. The \code{distance}
column gives the number of direct-parent hops from the input term to the
ancestor.
}
\description{
Find ancestor terms
}
