% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{mesh_disconnect}
\alias{mesh_disconnect}
\title{Disconnect a MeSH DuckDB connection}
\usage{
mesh_disconnect(db, shutdown = TRUE)
}
\arguments{
\item{db}{A \code{meshdb} object returned by \code{\link[=meshdb]{meshdb()}}.}

\item{shutdown}{If \code{TRUE}, shut down the DuckDB driver.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Disconnect a MeSH DuckDB connection
}
