% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mesh_map}
\alias{mesh_map}
\title{Query PubMed-to-MeSH mappings}
\usage{
mesh_map(
  species = c("human", "mouse"),
  pmid = NULL,
  mesh = NULL,
  gene = NULL,
  db = meshdb()
)
}
\arguments{
\item{species}{Either \code{"human"} or \code{"mouse"}.}

\item{pmid}{Optional PubMed IDs.}

\item{mesh}{Optional MeSH IDs or exact MeSH term labels.}

\item{gene}{Optional gene symbols.}

\item{db}{A \code{meshdb} object returned by \code{\link[=meshdb]{meshdb()}}.}
}
\value{
A tibble of matching PubMed-to-MeSH records.
}
\description{
Query PubMed-to-MeSH mappings
}
