% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationships.R
\name{mesh_parents}
\alias{mesh_parents}
\title{Find direct parent terms}
\usage{
mesh_parents(mesh, db = meshdb())
}
\arguments{
\item{mesh}{Character vector of MeSH IDs or exact MeSH term labels.}

\item{db}{A \code{meshdb} object returned by \code{\link[=meshdb]{meshdb()}}.}
}
\value{
A tibble with one row per input term and direct parent.
}
\description{
Find direct parent terms
}
