% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{mesh_url}
\alias{mesh_url}
\title{Construct a URL for a mesh term}
\usage{
mesh_url(x, term = NULL, ahref = FALSE, target = NULL, ...)
}
\arguments{
\item{x}{the mesh term (or vector of them)}
}
\description{
Construct a URL for a mesh term
}
