% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshdb-package.R
\docType{package}
\name{meshdb-package}
\alias{meshdb-package}
\title{meshdb: Navigate a Local MeSH Ontology Database}
\description{
Tools for querying MeSH terms and traversing parent-child and ancestor
relationships stored in local parquet files.
}
\author{
\strong{Maintainer}: Steve User \email{steve@example.com}

Authors:
\itemize{
  \item Steve User \email{steve@example.com}
}

}
\keyword{internal}
