% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationships.R
\name{mesh_children}
\alias{mesh_children}
\title{Find child terms}
\usage{
mesh_children(mesh, db = meshdb())
}
\arguments{
\item{mesh}{Character vector of MeSH IDs or exact MeSH term labels.}

\item{db}{A \code{meshdb} object returned by \code{\link[=meshdb]{meshdb()}}.}
}
\value{
A tibble with one row per input term and child. The \code{distance}
column gives the number of direct-child hops from the input term to the
child.
}
\description{
Find child terms
}
